pyproject.toml
[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"

[project]
name = "refua"
version = "0.0.1"
requires-python = ">=3.11,<3.15"
description = "Unified drug discovery ML toolkit"
readme = "README.md"
license = "MIT"
authors = [{name = "JJ Ben-Joseph", email = "jj@tensorspace.ai"}]
keywords = [
    "drug discovery",
    "machine learning",
    "protein design",
    "structure prediction",
    "generative design",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
dependencies = [
    "torch>=2.4.1",
    "numpy>=1.26,<2.0",
    "hydra-core>=1.3.2",
    "pytorch-lightning>=2.5.0",
    "rdkit>=2024.3.2",
    "dm-tree>=0.1.8",
    "requests>=2.32.3",
    "einops>=0.8.0",
    "einx>=0.3.0",
    "fairscale>=0.4.13",
    "mashumaro>=3.14",
    "modelcif>=1.2",
    "click>=8.1.7",
    "pyyaml>=6.0.2",
    "biopython>=1.84",
    "scipy>=1.13.1",
    "numba>=0.61.0",
    "gemmi>=0.6.5",
    "scikit-learn>=1.6.1",
    "chembl_structure_pipeline>=1.2.2",
    "biotite",
    "pydssp",
    "huggingface_hub",
    "pandas>=2.2.2",
    "matplotlib",
    "logomaker",
    "tqdm",
]

[project.urls]
Homepage = "https://tensorspace.ai/"
Repository = "https://github.com/tensorspace-ai/refua"

[project.scripts]
boltz = "refua.boltz.main:cli"
boltzgen = "refua.boltzgen.cli.boltzgen:main"

[project.optional-dependencies]
lint = ["ruff"]
test = ["pytest", "requests", "pandas>=2.2.2"]
doc = [
    "sphinx>=7",
    "furo>=2024.1.29",
    "myst-parser>=2.0.0",
    "sphinx-autodoc-typehints>=1.25.2",
]
cuda = [
    "cuequivariance_ops_cu12>=0.5.0",
    "cuequivariance_ops_torch_cu12>=0.5.0",
    "cuequivariance_torch>=0.5.0",
    "nvidia-ml-py>=12.535.133",
]
analysis = [
    "pandas>=2.2.2",
    "matplotlib",
    "logomaker",
    "hydride",
]
train = [
    "wandb>=0.18.7",
    "pandas>=2.2.2",
    "matplotlib",
]
process = [
    "pdbeccdutils",
    "p_tqdm",
    "redis",
]
dev = [
    "ruff",
    "pre-commit",
    "pytest",
    "requests",
    "types-requests>=2.32.4",
    "pandas>=2.2.2",
    "redis",
    "wandb>=0.18.7",
    "sphinx>=7",
    "furo>=2024.1.29",
    "myst-parser>=2.0.0",
    "sphinx-autodoc-typehints>=1.25.2",
]

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
refua.boltzgen = ["resources/**/*"]

[tool.ruff]
src = ["src"]
extend-exclude = ["conf.py"]
target-version = "py311"
lint.select = ["ALL"]
lint.ignore = [
    "COM812",  # Conflicts with the formatter
    "ISC001",  # Conflicts with the formatter
    "RET504",  # Unnecessary assignment to `x` before `return` statement
    "S101",    # Use of `assert` detected
    "D100",    # Missing docstring in public module
    "D104",    # Missing docstring in public package
    "PT001",   # https://github.com/astral-sh/ruff/issues/8796#issuecomment-1825907715
    "PT023",   # https://github.com/astral-sh/ruff/issues/8796#issuecomment-1825907715
    "FBT001",
    "FBT002",
    "PLR0913", # Too many arguments to init (> 5)
    "TRY003",  # Exception messages
    "EM101",   # Exception messages
    "FA102",   # typing
    "T201",    # Print statements
]

[tool.ruff.lint.per-file-ignores]
"**/__init__.py" = [
    "F401", # Imported but unused
    "F403", # Wildcard imports
]
"docs/**" = [
    "INP001", # Requires __init__.py but folder is not a package.
]
"scripts/**" = [
    "INP001", # Requires __init__.py but folder is not a package.
]
"tests/**" = [
    "S101", # Use of `assert` detected
    "D103", # Missing docstring in public function
]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`
keep-runtime-typing = true

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "regression",
]
pythonpath = ["src"]

[tool.mypy]
# Functions need to be annotated
disallow_untyped_defs = true
warn_unused_ignores = true
exclude = ["configs/", "build/", "dist/", "docs/", "tests/"]
